Fixed gcc compiler errors; particularly centroid calculation in mesh.cpp and initialition of variables (to =0) in bvh.cpp

+++ src/math/color.hpp
@@ -24,7 +24,7 @@ public:
     /**
      * The number of dimensions.
      */
-    static const size_t DIM = 3;
+    static const int DIM = 3;
 
     static const Color3 Black;
     static const Color3 White;
@@ -107,7 +107,7 @@ public:
     /**
      * @remark No bounds checking.
      */
-    const real_t& operator[]( size_t i ) const
+    const real_t& operator[]( int i ) const
     {
         // assumes all members are in a contiguous block
         assert( i >= 0 && i < DIM );
@@ -117,7 +117,7 @@ public:
     /**
      * @remark No bounds checking.
      */
-    real_t& operator[]( size_t i )
+    real_t& operator[]( int i )
     {
         // assumes all members are in a contiguous block
         assert( i >= 0 && i < DIM );
